Container like functionality and refactor

+++ Json4CPP/JsonDefault.cpp
@@ -6,5 +6,5 @@ using namespace std;
 
 namespace Json4CPP
 {
-  int JsonDefault::Indentation = 2;
+  uint8_t JsonDefault::Indentation = 2;
 }
